% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateAuc_Cal.R
\name{GenerateAuc_Cal}
\alias{GenerateAuc_Cal}
\title{Title}
\usage{
GenerateAuc_Cal(log.odds.standard, log.odds.intrinsic, log.odds.dic,
  log.odds.intrinsic.eb, log.odds.intrinsic.la, log.odds.intrinsic.tree, x.test,
  y.test)
}
\arguments{
\item{y.test}{}
}
\description{
Title
}
