% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCoeff.R
\name{GenerateCoeff}
\alias{GenerateCoeff}
\title{Title}
\usage{
GenerateCoeff(log.odds.standard, log.odds.intrinsic, log.odds.dic,
  log.odds.intrinsic.eb, log.odds.intrinsic.la, x.test, y.test)
}
\arguments{
\item{y.test}{}
}
\description{
Title
}
