% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_model_data_analysis.R
\name{two_data_two_stage_random}
\alias{two_data_two_stage_random}
\title{Title}
\usage{
two_data_two_stage_random(y.pheno.mis1, gene1, x.covar1, y.pheno.mis2, gene2,
  x.covar2)
}
\arguments{
\item{x.covar2}{}
}
\description{
Title
}
